gsplat-rust-renderer: simplify pass (Codex-reviewed)
From a Codex simplify-code review of the relocated apis/ + scene_io:

- log_gaussian_ply: complete the thin-shim refactor — move LogPlyConfig,
  splat_blueprint, and main into apis/log_gaussian_ply.py (now beartype-
  instrumented like the other four tools); tools/ is a ~7-line shim. Caught the
  same parents[1]->parents[2] path trap as relog_check: DEFAULT_PLY now resolves
  to the package-root examples/chair.ply from its deeper apis/ location.
- visualize_brush_training: extract ready_export_plys(), a generator that yields
  the unprocessed, size-stable export_*.ply checkpoints. Both watch loops
  duplicated the same walrus-glob scan + done_plys/awaiting_stable_size skips;
  the per-loop bodies (which differ) stay inline.
- visualize_brush_training: make the eval-disabled (eval_split_every == 0) path
  an explicit if/else instead of a buried comprehension guard.

Left alone per the review: scene_io grouping (cohesive), qvec_to_rotmat /
_colmap_fx_fy_cx_cy (small but encode COLMAP conventions), brush_blueprint's
nested cell helpers (keep a dense blueprint readable), and calibration_scene's
in-package Tyro subcommand parse (moving it to the shim would drop it out of
beartype instrumentation).

Verified: ruff + pyrefly clean, 13 tests, all 5 CLIs --help, DEFAULT_PLY resolves.
